Add background jobs to create and update trips
A `Trip` joins a `Ride` and a `Driver` and is created for every
`Ride`/`Driver` combination.  The `Trip` is updated when the relevant
attributes of the `Ride` or `Driver` change.

Each `Trip` has a `score` attribute that is calculated from the
metrics populated by the `AnalyzeTripJob`, which calls the Google
Directions API.