Add seed data
Running the `rails db:seed` command will create drivers and rides
using real addresses (so that the Google Directions API will
produce usable results).

Note that `Trips` are created by callbacks in the `Ride` and `Driver`
models, which enqueue background jobs to create the trips and then call
the Google Directions API to populate trip metrics. I chose an async
approach so that the database would be seeded with real data, and so
that (in the future) creating a `Ride` or `Driver` won't be blocked in
the UI by numerous API calls.

Running `bundle exec sidekiq` will start the background jobs after
the seed data is created.

```ruby